+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - it's now possible to set background layer with manual settings
 - add an edit button in the data browser (when in edit mode)
 - add icon URL formatting with feature properties
+- add "Transform to Polygon/Polyline" action
 
 ## 0.4.x
 - add a data browser